% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESpace_test.R
\name{DESpace_test}
\alias{DESpace_test}
\title{DESpace_test}
\usage{
DESpace_test(
  spe,
  spatial_cluster,
  sample_col = NULL,
  replicates = FALSE,
  min_counts = 20,
  min_non_zero_spots = 10,
  filter_gene = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{spe}{SpatialExperiment or SingleCellExperiment.}

\item{spatial_cluster}{Column name of spatial clusters in \code{colData(spe)}.}

\item{sample_col}{Column name of sample ids in \code{colData(spe)}.}

\item{replicates}{A logical, indicating whether biological replicates are provided (TRUE) or not (FALSE).
If biological replicates are provided, \code{\link{DESpace_test}} performs a joint test across all replicates, 
searching for SVGs with consistent spatial patterns across samples.}

\item{min_counts}{Minimum number of counts per sample (across all spots) for a gene to be analyzed.}

\item{min_non_zero_spots}{Minimum number of non-zero spots per sample, for a gene to be analyzed.}

\item{filter_gene}{A logical. If TRUE, 
\code{\link{DESpace_test}} filters genes:
genes have to be expressed in at least 'min_non_zero_spots' spots, 
and a gene requires at least 'min counts' counts per sample (across all locations).}

\item{verbose}{A logical.
If TRUE, \code{\link{DESpace_test}} returns two more results: 
'DGEGLM' and 'DGELRT' objects contain full statistics from 'edgeR::glmFit' and 'edgeR::glmLRT'.}
}
\value{
A list of results:

- "gene_results": a dataframe contains main edgeR test results;

- "estimated_y": a DGEList object contains the estimated common dispersion, 
which can later be used to speed-up calculation when testing individual clusters.

- "glmFit" (only if \code{verbose = TRUE}): a DGEGLM object contains full statistics from "edgeR::glmFit".

- "glmLRT" (only if \code{verbose = TRUE}): a DGELRT object contains full statistics from "edgeR::glmLRT".
}
\description{
'DESpace_test' identifies spatially variable genes (SVGs) from spatially-resolved transcriptomics data,
provided spatial clusters are available.
}
\details{
If 'sample_col' is not specified and 'replicates == FALSE',
\code{\link{DESpace_test}} assumed that data comes from an individual sample, 
and performs SV testing on it.

If 'sample_col' is provided and 'replicates == FALSE',
\code{\link{DESpace_test}} tests each sample individually and returns a list of results for each sample.

If 'sample_col' is provided and 'replicates == TRUE',
\code{\link{DESpace_test}} performs a joint multi-sample test.
}
\examples{
# load the package
# Connect to ExperimentHub
ehub <- ExperimentHub::ExperimentHub()
# Download the example spe data
spe_all <- spatialLIBD::fetch_data(type = "spe", eh = ehub)
spe_all

# Only use one sample:
library(SpatialExperiment)
spe3 <- spe_all[, colData(spe_all)$sample_id == '151673']
rm(spe_all)
# Select small set of random genes for faster runtime in this example
set.seed(123)
sel_genes <- sample(dim(spe3)[1],500)
spe3 <- spe3[sel_genes,]
# Discard lowly abundant genes
qc_low_gene <- rowSums(assays(spe3)$counts > 0) >=10
# Remove lowly abundant genes
spe3 = spe3[qc_low_gene,]

# Fit the model via \code{\link{DESpace_test}} function.
set.seed(123)
results_DESpace <- DESpace_test(spe = spe3,
                           spatial_cluster = "layer_guess_reordered",
                           verbose = FALSE)

# DESpace_test returns of a list of 2 objects:
# "gene_results": a dataframe contains main edgeR test results;
# "estimated_y": a DGEList object contains the estimated common dispersion, 
#  which can later be used to speed-up calculation when testing individual clusters.

# We visualize differential results:
head(results_DESpace$gene_results, 3)

}
\seealso{
\code{\link{top_results}}, \code{\link{individual_test}}, \code{\link{FeaturePlot}}
}
