% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturePlot.R
\name{FeaturePlot}
\alias{FeaturePlot}
\title{FeaturePlot}
\usage{
FeaturePlot(
  spe,
  feature,
  coordinates = NULL,
  assay.type = "logcounts",
  Annotated_cluster = FALSE,
  diverging = FALSE,
  low = NULL,
  high = NULL,
  mid = NULL,
  color = NULL,
  platform = "Visium",
  spatial_cluster = NULL,
  cluster = NULL,
  legend_layer = FALSE,
  label = FALSE,
  ncol = 3,
  title = FALSE,
  linewidth = 0.4,
  legend_exprs = FALSE,
  title_size = 5
)
}
\arguments{
\item{spe}{SpatialExperiment or SingleCellExperiment. If \code{feature} is specified and is a 
string, it must exist as a row in the specified assay of \code{spe}.}

\item{feature}{Feature vector used to color each spot. May be the name of a
gene/row in an assay of \code{spe}, or a vector of continuous values.}

\item{coordinates}{Column names of spatial coordinates of spots stored in \code{colData(spe)}.}

\item{assay.type}{String indicating which assay in \code{spe} the expression
vector should be taken from.}

\item{Annotated_cluster}{A logical. TRUE or FALSE, indicating whether to plot the annotated spatial clusters next to expression plots.}

\item{diverging}{A logical. If true, use a diverging color gradient in
\code{\link{FeaturePlot}} (e.g. when plotting a fold change) instead of a
sequential gradient (e.g. when plotting expression).}

\item{low, mid, high}{Optional hex codes for low, mid, and high values of the
color gradient used for continuous spot values.}

\item{color}{Optional hex code to set color of borders around spots. Set to
\code{NA} to remove borders.}

\item{platform}{Spatial sequencing platform. If "Visium", the hex spot layout
will be used, otherwise square spots will be plotted.\cr
NOTE: specifying this argument is only necessary if \code{spe} was not
created by \code{BayesSpace::patialCluster()} or \code{BayesSpace::spatialEnhance()}.}

\item{spatial_cluster}{Column name of spatial clusters in \code{colData(spe)}.}

\item{cluster}{Names of the spatial clusters used for drawing a boundary 
around a group of points that belong to the specify cluster.
It can be NULL, "all"/"ALL", or a vector of cluster names.}

\item{legend_layer}{A logical. TRUE of FALSE, indicating whether to plot the legend for the shaped layers (TRUE), or not (FALSE).
Only used when 'spatial_cluster' and 'cluster' are specified.}

\item{label}{A logical. TRUE of FALSE. Adding a label and an arrow pointing to a group.}

\item{ncol}{The dimensions of the grid to create. 
By default, 1, if the length of feature equals to 1, and 3, otherwise.}

\item{title}{A logical. TRUE or FALSE. If true, the title name of each (subplot) is the gene name.}

\item{linewidth}{The width of the boundary line around the cluster.
The default ('0.4') size of the boundary line is one.}

\item{legend_exprs}{A logical. TRUE of FALSE, indicating whether to plot the legend for the expression level (TRUE), or not (FALSE).}

\item{title_size}{Text size.}
}
\value{
Returns a ggplot object.
}
\description{
Plot spatial gene expression.
This function is a modified version of the \code{\link{FeaturePlot}} function from BayesSpace R package.
In comparison to the original BayesSpace function, this function allows plotting multiple genes simultaneously and drawing an outline around a specified cluster.
}
\examples{
# load the input data:
data("LIBD_subset", package = "DESpace")
LIBD_subset

# load pre-computed results (obtained via `DESpace_test`)
data("results_DESpace_test", package = "DESpace")

# DESpace_test returns of a list of 2 objects:
# "gene_results": a dataframe contains main edgeR test results;
# "estimated_y": a DGEList object contains the estimated common dispersion, 
#  which can later be used to speed-up calculation when testing individual clusters.

# We visualize differential results:
head(results_DESpace_test$gene_results, 3)

# Visualize the gene expression of the top three genes
feature = results_DESpace_test$gene_results$gene_id[seq_len(3)]
feature
library(SpatialExperiment)
FeaturePlot(LIBD_subset, feature, coordinates = c("array_row", "array_col"),
            ncol = 3, title = TRUE)

}
\seealso{
\code{\link{DESpace_test}}, \code{\link{individual_test}}, \code{\link{top_results}}
}
\concept{spatial plotting functions}
