% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_test.R
\name{individual_test}
\alias{individual_test}
\title{individual_test}
\usage{
individual_test(
  spe,
  spatial_cluster,
  sample_col = "sample_id",
  edgeR_y = NULL,
  min_counts = 20,
  min_non_zero_spots = 10,
  filter_gene = TRUE,
  replicates = FALSE,
  BPPARAM = NULL
)
}
\arguments{
\item{spe}{SpatialExperiment or SingleCellExperiment.}

\item{spatial_cluster}{Column name of spatial clusters in \code{colData(spe)}.}

\item{sample_col}{Column name of sample ids in \code{colData(spe)}.}

\item{edgeR_y}{Pre-estimated dispersion; if it's null, compute dispersion.}

\item{min_counts}{Minimum number of counts per sample (across all spots) for a gene to be analyzed.}

\item{min_non_zero_spots}{Minimum number of non-zero spots per sample, for a gene to be analyzed.}

\item{filter_gene}{A logical. If TRUE, 
\code{\link{DESpace_test}} filters genes:
genes have to be expressed in at least 'min_non_zero_spots' spots, 
and a gene requires at least 'min counts' counts per sample (across all locations).}

\item{replicates}{Single sample or multi-sample test.}

\item{BPPARAM}{An optional parameter passed internally to bplapply.
We suggest using as many cores as the number of spatial clusters.
If unspecified, the script does not run in parallel.
Note that parallelizing the script will increase the memory requirement;
if memory is an issue, leave 'BPPARAM' unspecified and, hence, avoid parallelization.}
}
\value{
List of results (gene_results, estimated_y, glmLrt and glmFit): 
  edgeR test results, estimated dispersion, full statistics from 'edgeR::glmFit' and 'edgeR::glmLRT'.
}
\description{
DESpace can also be used to reveal the specific areas of the tissue affected by SVGs; i.e., spatial clusters that are particularly over/under abundant compared to the average signal.
This function can be used to identify SVGs for each individual cluster.
}
\details{
For every spatial cluster we test, \code{edgeR} would normally re-compute the dispersion estimates based on the specific design of the test.
However, this calculation represents the majority of the overall computing time.
Therefore, to speed-up calculations, we propose to use the dispersion estimates which were previously computed for the gene-level tests.
This introduces a minor approximation which, in our benchmarks, does not lead to decreased accuracy.
If you want to use pre-computed gene-level dispersion estimates, set \code{edgeR_y} to 'estimated_y'.
Alternatively, if you want to re-compute dispersion estimates (significantly slower, but marginally more accurate option), leave edgeR_y empty.
}
\examples{
# Connect to ExperimentHub
ehub <- ExperimentHub::ExperimentHub()
# Download the example spe data
spe_all <- spatialLIBD::fetch_data(type = "spe", eh = ehub)
spe_all

# Only use one sample:
library(SpatialExperiment)
spe3 <- spe_all[, colData(spe_all)$sample_id == '151673']
rm(spe_all)
# Select small set of random genes for faster runtime in this example
set.seed(123)
sel_genes <- sample(dim(spe3)[1],500)
spe3 <- spe3[sel_genes,]

# load pre-computed results (obtaines via `DESpace_test`)
data("results_DESpace_test", package = "DESpace")

# DESpace_test returns of a list of 2 objects:
# "gene_results": a dataframe contains main edgeR test results;
# "estimated_y": a DGEList object contains the estimated common dispersion, 
#  which can later be used to speed-up calculation when testing individual clusters.

# We visualize differential results:
head(results_DESpace_test$gene_results, 3)

# Individual cluster test: identify SVGs for each individual cluster
# set parallel computing; we suggest using as many cores as the number of spatial clusters.
# Note that parallelizing the script will increase the memory requirement;
# if memory is an issue, leave 'BPPARAM' unspecified and, hence, avoid parallelization.
BPPARAM = BiocParallel::SnowParam(workers = 2, RNGseed = 100)
set.seed(123)
results_individual_test <- individual_test(spe3, 
                                           edgeR_y = results_DESpace_test$estimated_y, 
                                           spatial_cluster = "layer_guess_reordered",
                                           BPPARAM = BPPARAM)
                                           
# We visualize results for the cluster 'WM'
results_WM <- results_individual_test[[7]]
head(results_WM,3)

}
\seealso{
\code{\link{top_results}}, \code{\link{DESpace_test}}, \code{\link{FeaturePlot}}
}
