% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_LIBD_multi.R
\name{LIBD_multi}
\alias{LIBD_multi}
\title{Subset from the human DLPFC 10x Genomics Visium dataset of the \code{spatialLIBD} package}
\arguments{
\item{LIBD_multicontains}{a \code{\linkS4class{SpatialExperiment}} object,
representing a subset of the sample 151507, 151509, 151673, and 151675 from the full real data of the \code{spatialLIBD} package.
Below the code used to subset the original dataset.}
}
\description{
Subset from the human DLPFC 10x Genomics Visium dataset of the \code{spatialLIBD} package
}
\examples{
# Connect to ExperimentHub
# ehub <- ExperimentHub::ExperimentHub()
# Download the example spe data
# spe_all <- spatialLIBD::fetch_data(type = "spe", eh = ehub)
# Select one sample only:
# LIBD_multi <- spe_all[, colData(spe_all)$sample_id \%in\% c('151507', '151509', '151673', '151675')]
# Select small set of random genes for faster runtime 
# set.seed(123)
# sel_genes <- sample(dim(LIBD_multi)[1],100)
# LIBD_multi <- LIBD_multi[sel_genes,]
# keep_col <- c("array_row","array_col","layer_guess_reordered", "sample_id")
# library(SpatialExperiment)
# LIBD_multi <- SpatialExperiment(assay = list(counts = assay(LIBD_multi),
#                                               logcounts = logcounts(LIBD_multi)), 
#                                  colData = colData(LIBD_multi)[keep_col])
# colData(LIBD_multi)[["condition"]] <- ifelse(colData(LIBD_multi)[["sample_id"]] \%in\% c("151507", "151509"), "1", "2")
# save(LIBD_multi, file = "./DESpace/data/LIBD_multi.RData")

}
\seealso{
\code{\link{dsp_test}}, \code{\link{individual_dsp}},
}
\author{
Peiying Cai \email{peiying.cai@uzh.ch}, Simone Tiberi \email{simone.tiberi@unibo.it}
}
