% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_results_individual_svg.R
\name{results_individual_svg}
\alias{results_individual_svg}
\title{Results from \code{\link{individual_svg}} function}
\arguments{
\item{results_individual_svg}{contains a \code{\linkS4class{list}} object,
with the results obtained applying \code{\link{individual_svg}} function to an external dataset from the spatialLIBD package.
Below the code used to obtain 'results_individual_svg'.}
}
\value{
A List of 7 elements - one element for each spatial cluster
}
\description{
Results from \code{\link{individual_svg}} function
}
\examples{
# load the input data:
# data("LIBD_subset", package = "DESpace")
# LIBD_subset
# load pre-computed results (obtained via `DESpace_test`)
# data("results_DESpace_test", package = "DESpace")
# results_DESpace_test

# Function `individual_svg()` can be used to identify SVGs for each individual cluster.
# Parameter `spatial_cluster` indicates the column names of `colData(spe)` 
# containing spatial clusters.
# set.seed(123)
# results_individual_svg <- individual_svg(LIBD_subset,
#                                            edgeR_y = results_DESpace_test$estimated_y,
#                                            spatial_cluster = "layer_guess_reordered")
# save(results_individual_svg, file = "./DESpace/data/results_individual_svg.RData")
}
\seealso{
\code{\link{svg_test}}, \code{\link{individual_svg}}
}
\author{
Peiying Cai \email{peiying.cai@uzh.ch}, Simone Tiberi \email{simone.tiberi@unibo.it}
}
